+++ app/models/device.rb
@@ -1,0 +1,4 @@
+class Device < ApplicationRecord
+  belongs_to :client
+  belongs_to :site, optional: true
+end