+++ app/models/site.rb
@@ -1,3 +1,5 @@
 class Site < ApplicationRecord
   belongs_to :tenant
+  has_many :devices, dependent: :destroy
+  has_many :clients, through: :devices
 end